add shell tool timeout parameter and clean up tool headers
Shell gains an optional "timeout" (seconds) parameter that is
threaded through to exec() so the model can bound how long a
long-ish command is allowed to run (default remains 10s).

While touching the tool headers, switch the #include for
agt/tool.hpp to angle-bracket form, apply consistent pointer/
reference spacing, and add the explicit size_t / streamsize casts
FileRead / FileWrite need on 64-bit builds. Minor README wording
tweak tagging along.

+++ src/file_read.hpp
@@ -1,7 +1,8 @@
 #pragma once
 
-#include "agt/tool.hpp"
+#include <agt/tool.hpp>
 #include "exec.hpp"
+#include <cstddef>
 #include <filesystem>
 #include <fstream>
 #include <iterator>
@@ -35,9 +36,9 @@ class FileRead : public agt::Tool {
 
     if (input.contains("len")) {
       auto len = input["len"].get<long>();
-      std::string buf(len, '\0');
+      std::string buf(static_cast<std::size_t>(len), '\0');
       in.read(buf.data(), len);
-      buf.resize(in.gcount());
+      buf.resize(static_cast<std::size_t>(in.gcount()));
       return buf;
     }
